move image pull to its own file
